backend/requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1

# Database
psycopg[binary]>=3.1.0
dj-database-url>=2.1.0

# Task Queue
celery==5.3.4
redis==5.0.1
daphne==4.1.0
channels==4.0.0
channels-redis==4.2.0

# Environment Variables
python-decouple==3.8

# Plaid Integration
plaid-python>=20.0.0

# Stripe Integration
stripe>=10.0.0

# Security
cryptography==41.0.7
pyotp==2.9.0
qrcode==7.4.2

# API Documentation
drf-yasg==1.21.7

# Testing
pytest-django==4.7.0
factory-boy==3.3.0
pytest-cov==4.1.0

# Development Tools
django-debug-toolbar>=4.0.0

# Storage
django-storages>=1.14.0
boto3>=1.34.0

# Utilities
python-dateutil==2.8.2
django-filter==23.5
requests==2.31.0
setuptools>=80.0.0
psutil>=5.9.0

